refactor: keep DAG state record encoding with persistence
